% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBFH.R
\name{EmpiricalBayes}
\alias{EmpiricalBayes}
\title{Title}
\usage{
EmpiricalBayes(response, designmatrix, sampling.var)
}
\arguments{
\item{response}{a numeric vector. It represents the response or the observed value in the Fay Herriot Model}

\item{designmatrix}{a numeric matrix. The first column is a column of ones(also called the intercept). The other columns consist of observations of each of the covariates or the explanatory variable in Fay Herriot Model.}

\item{sampling.var}{a numeric vector consisting of the known sampling variances of each of the small area levels.}
}
\value{
estimate
}
\description{
Title
}
\examples{
1
}
