% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Jackknife.R
\name{mspe_Jackknife}
\alias{mspe_Jackknife}
\title{Title}
\usage{
mspe_Jackknife(response, designmatrix, sampling.var, method = "REML")
}
\arguments{
\item{response}{a numeric vector. It represents the response or the observed value in the Fay Herriot Model}

\item{designmatrix}{a numeric matrix. The first column is a column of ones(also called the intercept). The other columns consist of observations of each of the covariates or the explanatory variable in Fay Herriot Model.}

\item{sampling.var}{a numeric vector consisting of the known sampling variances of each of the small area levels.}

\item{method}{Ahat estimator method}
}
\value{
1
}
\description{
Title
}
\examples{
1
}
