% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NER_DL.R
\name{mspe_NER_DL}
\alias{mspe_NER_DL}
\title{Title}
\usage{
mspe_NER_DL(ni, X, Y, X.mean, x.bar, y.bar, method = "REML")
}
\arguments{
\item{ni}{sampled units observed in the ith small area}

\item{X}{auxiliary values matrix}

\item{Y}{response variable}

\item{X.mean}{mean auxiliary value of population}

\item{x.bar}{mean auxiliary value of small area}

\item{y.bar}{mean response value of small area}

\item{method}{estimation method of sig_e2 and sig_v2}
}
\value{
mspe of NER model using DL method
}
\description{
Title
}
\examples{
1
}
