% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NER_PR.R
\name{mspe_NER_PR}
\alias{mspe_NER_PR}
\title{Title}
\usage{
mspe_NER_PR(ni, X, Y, X.mean, x.bar, y.bar)
}
\arguments{
\item{ni}{sampled units observed in the ith small area}

\item{X}{auxiliary values matrix}

\item{Y}{response variable}

\item{X.mean}{mean auxiliary value of population}

\item{x.bar}{mean auxiliary value of small area}

\item{y.bar}{mean response value of small area}
}
\value{
mspe of NER model using PR method
}
\description{
Title
}
\examples{
1
}
