% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_reml.R
\name{NER_REML}
\alias{NER_REML}
\title{Title}
\usage{
NER_REML(p, XY, x.bar, y.bar, K = 1000)
}
\arguments{
\item{p}{dimension of auxiliary matrix X}

\item{XY}{data matrix containing X,Y and its area number}

\item{x.bar}{mean auxiliary value of small area}

\item{y.bar}{mean response value of small area}

\item{K}{searching number}
}
\value{
resimaxiliklihood estimation of NER model and its blup beta
}
\description{
Title
}
\examples{
1
}
