\name{saeMSPE-package}
\alias{saeMSPE-package}
\alias{saeMSPE}
\docType{package}
\title{
\packageTitle{saeMSPE}
}
\description{
\packageDescription{saeMSPE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{saeMSPE}

}
\author{
\packageAuthor{saeMSPE}

Maintainer: \packageMaintainer{saeMSPE}
}
\references{
J. Jiang and M. Torabi. Sumca: simple; unified; monte carlo assisted approach to second order unbiased mean squared prediction error estimation. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 82(2):467–485, 2020
}

\keyword{ package }

\examples{
1
}
