% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdataFH.R
\name{sim_data_FH}
\alias{sim_data_FH}
\title{Title}
\usage{
sim_data_FH(beta, A, D, m, p)
}
\arguments{
\item{beta}{coefficient}

\item{A}{unknown sampling variance}

\item{D}{known sampling variance}

\item{m}{sampling number}

\item{p}{dimention of sample}
}
\value{
simulation data for Fay Herriot model
}
\description{
Title
}
\examples{
1
}
