\name{smaFHNER-package}
\alias{smaFHNER-package}
\alias{smaFHNER}
\docType{package}
\title{
\packageTitle{smaFHNER}
}
\description{
\packageDescription{smaFHNER}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{smaFHNER}
}
\author{
\packageAuthor{smaFHNER}

Maintainer: \packageMaintainer{smaFHNER}
}
\references{
Prasad, N.G.N. , and Rao, J.N.K.(1990). The estimation of the mean squared error of small-area
estimators. Publications of the American Statistical Association, 85(409), 163-171.
}

\keyword{ package }

\examples{
1
}
